fix bug

+++ app/src/main/java/org/cxct/sportlottery/ui/money/recharge/OnlinePayFragment.kt
@@ -525,7 +525,7 @@ class OnlinePayFragment : BaseFragment<MoneyRechViewModel, OnlinePayFragmentBind
         val additional = dailyConfig.additional
         val capped = dailyConfig.capped
         if (additional > 0) {
-            val additionalMoney = rechargeMoney*additional/100
+            val additionalMoney = rechargeMoney.toDouble() * additional / 100
             val extraMoney = if(additionalMoney>capped) capped else  additionalMoney
             binding.tvExtraAmount.text = "${sConfigData?.systemCurrencySign} ${TextUtil.formatMoney(extraMoney,2)}"
         }
